feat: emit standalone scalar support

+++ testdata/projects/demand-control/api/expected.ts
@@ -1,9 +1,9 @@
-import type { int64 } from "@tsonic/core/types.js";
+import type { int32 } from "../../../support/scalars.js";
 import { Sum as Sum__from_example__u2e_com__u2f_control__u2f_worker } from "../worker/worker.js";
-export function Run(value: int64): int64 {
+export function Run(value: int32): int32 {
     switch (value) {
-        case 0 as int64: {
-            return 0 as int64;
+        case 0 as int32: {
+            return 0 as int32;
             break;
         }
         default: {
