+++ testdata/projects/void-calls/expected.ts
@@ -1,13 +1,13 @@
-import type { int64 } from "@tsonic/core/types.js";
-export function Touch(value: int64): void {
-    if (value > (0 as int64)) {
+import type { int32 } from "../../../support/scalars.js";
+export function Touch(value: int32): void {
+    if (value > (0 as int32)) {
         return;
     }
 }
-export function Identity(value: int64): int64 {
+export function Identity(value: int32): int32 {
     return value;
 }
-export function Run(value: int64): int64 {
+export function Run(value: int32): int32 {
     Touch(value);
     Identity(value);
     return value;
